+++ http-client/ChangeLog.md
@@ -1,3 +1,7 @@
+## next
+
+* Omit port for `getUri` when protocol is `https` and port is `443`
+
 ## 0.5.13.1
 
 * Add a workaround for a cabal bug [haskell-infra/hackage-trustees#165](https://github.com/haskell-infra/hackage-trustees/issues/165)